Create README.md
the program includes a thorough database of material attributes, including fatigue and fracture data, to allow for accurate material selection and fatigue life estimates. This information improves the design’s reliability and safety, lowering the chance of early failure or unanticipated performance difficulties. As well as, lowering the chance for human’s errors that can happen due to the constants that are used in the mathematical formulae needed to calculate & consider for the dynamic loads.